Extract navbar to components
This will simplify the root App.vue by moving the navbar logic out of it. That logic will was distributed to where it was needed inside of the component chain.
